+++ runtime/service/plugins/autotag/autotagplugin.h
@@ -32,7 +32,7 @@ namespace Nepomuk
         public:
             AutotagPlugin(QObject*, const QList<QVariant>&);
             virtual Nepomuk::WebExtractor::DataPP * getDataPP(KSharedConfigPtr configFile);
-            static float  version();
+            static int  version();
     };
 }
 #endif
